+++ wgpu/src/backend/webgpu.rs
@@ -1083,6 +1083,7 @@ fn future_request_device(
                     inner: device,
                     ident: crate::cmp::Identifier::create(),
                     error_scope_count: Rc::new(Cell::new(0)),
+                    uncaptured_error_listener: Rc::new(RefCell::new(None)),
                 }
                 .into(),
                 WebQueue {
@@ -1310,6 +1311,9 @@ pub struct WebDevice {
     ident: crate::cmp::Identifier,
     /// Current number of error scopes that have been pushed on the device.
     error_scope_count: Rc<Cell<u32>>,
+    /// The `uncapturederror` listener currently registered on the device, so
+    /// that setting a new handler can unregister the previous one.
+    uncaptured_error_listener: Rc<RefCell<Option<js_sys::Function>>>,
 }
 
 #[derive(Debug, Clone)]
@@ -2824,8 +2828,27 @@ impl dispatch::DeviceInterface for WebDevice {
             let error = error_from_js(event.error().value_of());
             handler(error);
         }) as Box<dyn FnMut(_)>);
+        let listener: js_sys::Function = f.as_ref().unchecked_ref::<js_sys::Function>().clone();
+
+        // Not every browser implements the `onuncapturederror` attribute
+        // (Safari does not, see <https://bugs.webkit.org/show_bug.cgi?id=323544>),
+        // where assigning to it silently does nothing. The event is dispatched
+        // either way, so listen for it instead.
         self.inner
-            .set_onuncapturederror(Some(f.as_ref().unchecked_ref()));
+            .add_event_listener_with_callback("uncapturederror", &listener)
+            .expect("Adding an event listener should never fail");
+
+        // Setting a handler replaces the previous one, as on the other backends.
+        let previous = self
+            .uncaptured_error_listener
+            .borrow_mut()
+            .replace(listener);
+        if let Some(previous) = previous {
+            self.inner
+                .remove_event_listener_with_callback("uncapturederror", &previous)
+                .expect("Removing an event listener should never fail");
+        }
+
         // Release memory management of this closure from Rust to the JS GC.
         // TODO: This will leak if weak references is not supported.
         f.forget();